requirements.txt
fastapi~=0.103.2
uvicorn~=0.23.2
pillow~=10.1.0
sqlalchemy~=2.0.22
python-dotenv~=1.0.0
alembic~=1.12.1
pydantic~=2.4.2
psycopg2-binary==2.9.6
fastapi-users[sqlalchemy]==12.1.2
asyncpg==0.29.0